refactor(context_compressor): extract skill-prune/ghost-skill cluster (LB2)
Wave-2 redraw: banner-inclusive window 404-618 (golden
ec5defbef6a402ad302b28369c81f5f9182162f1c8353b231d2d863fe59305c0) into
context_compressor_skill_prune.py. Deps on the two text-util leaves now
import one-way from context_compressor_text_utils — no cycle. 12 members
re-exported is-identical through the godfile. 35/35 targeted tests green.

Part of #78645
Part of #78647
